+++ CHANGELOG.md
@@ -6,6 +6,7 @@ GAMSPy 1.3.0
 - General
   - Change the way to show limited variables in latex file.
   - Overload __rpower__ for operables.
+  - Add a verification step for working directory path length.
 - Documentation
   - Add a section for limited variables. 
   - Add an example that shows how to read from another Container.